requirements.txt
Scrapy
SQLAlchemy
django
djangorestframework
psycopg2
bsddb3
scrapy-deltafetch
